// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/workflowDefinitionDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程
 * OpenAPI spec version: v1.0.0
 */
import type { WorkflowEngineBindingDTO } from "./workflowEngineBindingDTO";
import type { WorkflowDefinitionDTOPlatformType } from "./workflowDefinitionDTOPlatformType";
import type { WorkflowStageDTO } from "./workflowStageDTO";

export interface WorkflowDefinitionDTO {
  /** No comments found. */
  description?: string;
  /** No comments found. */
  enabled?: boolean;
  /** No comments found.(object) */
  engineBinding?: WorkflowEngineBindingDTO;
  /** No comments found.<br/>(See: 通用 ETL 平台类型。) */
  platformType?: WorkflowDefinitionDTOPlatformType;
  /** No comments found. */
  stages?: WorkflowStageDTO[];
  /** No comments found. */
  workflowCode?: string;
  /** No comments found. */
  workflowName?: string;
  /** No comments found. */
  workflowVersionNo?: number;
}